// Frontend/SpaceMatch/src/types/Booking.ts
export interface Booking {
    _id?: string; // Optional, will be generated by the database
    spaceId: string; // ID of the booked space
    dateOfBooking: Date; // Date when the booking was made
    startTime: Date; // Start time of the booking
    endTime: Date; // End time of the booking
    totalPrice: number; // Total price for the booking
    status?: 'confirmed' | 'payment outstanding' | 'cancelled'; // Booking status
    createdAt?: Date; // Creation date, optional
    updatedAt?: Date; // Last update date, optional
    bookingUserName?: string; // Username of the user who made the booking
}